+++ internal/store/postgres/queries/projects.sql
@@ -1,2 +1,7 @@
 -- name: GetProjectBySlug :one
 select * from projects where slug = $1;
+
+-- name: CreateProject :one
+insert into projects (name, slug)
+values ($1, $2)
+returning *;